requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
psycopg2
python-dateutil